Crate maud

A macro for writing HTML templates.

This documentation only describes the runtime API. For a general guide, check out the book instead.

Structs

Escaper
An adapter that escapes HTML special characters.

PreEscaped
A wrapper that renders the inner value without escaping.

Utf8Writer
Wraps a std::io::Write in a std::fmt::Write.

Traits

Render
Represents a type that can be rendered as HTML.